Removed Unused Interfaces also Dependencies From Controllers
