+++ API/Controllers/AddressController.cs
@@ -1,7 +1,7 @@
 using AutoMapper;
+using Contracts.DbContexts;
 using Contracts.DTOs.Address;
 using Domain.Entities.Address;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 [Route("api/[controller]")]
@@ -9,20 +9,19 @@ using Microsoft.AspNetCore.Mvc;
 public class AddressController : ControllerBase
 {
   private readonly IMapper _mapper;
-  private readonly IRepositoryManager _repoManager;
+  private readonly IAddressDbContext _addressDbContext;
 
+    public AddressController(IMapper mapper, IAddressDbContext addressDbContext)
+    {
+        _mapper = mapper;
+        _addressDbContext = addressDbContext;
+    }
 
-  public AddressController(IRepositoryManager repoManager, IMapper mapper)
-  {
-    _mapper = mapper;
-    _repoManager = repoManager;
-  }
-
-  [HttpPost, Route("new")]
+    [HttpPost, Route("new")]
   public async Task<ActionResult<Address?>> New(AddressCreateDTO request)
   {
-    byte[]? uid = _repoManager.Address.GetUserID(request.Mail);
-    uint? RegionID = _repoManager.Address.CheckRegionExist(request.RegionName);
+    byte[]? uid = _addressDbContext.GetUserID(request.Mail);
+    uint? RegionID = _addressDbContext.CheckRegionExist(request.RegionName);
 
     if (uid is null)
       return NotFound("User doesn't exist");
@@ -38,7 +37,7 @@ public class AddressController : ControllerBase
     Address.ID = AddressID;
     Address.RegionID = (uint)RegionID;
 
-    Address? Result = await _repoManager.Address.AddAddress(Address, uid);
+    Address? Result = await _addressDbContext.AddAddress(Address, uid);
 
     if (Result is not null)
       return Ok(_mapper.Map<Address, AddressPresentationDTO>(Result));
@@ -50,14 +49,14 @@ public class AddressController : ControllerBase
   [HttpPost, Route("modify")]
   public async Task<ActionResult<AddressPresentationDTO>> Modify(AddressPresentationDTO request)
   {
-    Address? Address = await _repoManager.Address.GetAddressByID(request.AddressID);
+    Address? Address = await _addressDbContext.GetAddressByID(request.AddressID);
     if (Address is null)
       return NotFound("Wrong addressID");
 
     var z = _mapper.Map<AddressPresentationDTO, Address>(request);
 
 
-    await _repoManager.Address.ModifyAddress(z, Address);
+    await _addressDbContext.ModifyAddress(z, Address);
 
     return Ok();
   }
@@ -65,11 +64,11 @@ public class AddressController : ControllerBase
   [HttpGet, Route("all")]
   public async Task<ActionResult<IEnumerable<AddressPresentationDTO>>> GetAll(string email)
   {
-    byte[]? uid = _repoManager.Address.GetUserID(email);
+    byte[]? uid = _addressDbContext.GetUserID(email);
     if (uid is null)
       return NotFound();
 
-    IEnumerable<Address>? addressList = await _repoManager.Address.GetUserAddresses(uid!);
+    IEnumerable<Address>? addressList = await _addressDbContext.GetUserAddresses(uid!);
 
     return Ok(_mapper.Map<IEnumerable<Address>, IEnumerable<AddressPresentationDTO>>(addressList!));
 
@@ -78,17 +77,17 @@ public class AddressController : ControllerBase
   [HttpGet, Route("regions")]
   public async Task<ActionResult<IEnumerable<Region>>> GetRegions()
   {
-    return Ok(await _repoManager.Address.GetRegions());
+    return Ok(await _addressDbContext.GetRegions());
   }
 
   [HttpPost, Route("delete")]
   public async Task<ActionResult> Delete(AddressDeleteDTO request)
   {
-    Address? Address = await _repoManager.Address.GetAddressByID(request.AddressID);
+    Address? Address = await _addressDbContext.GetAddressByID(request.AddressID);
     if (Address is null)
       return NotFound("Invalid AddressID!!");
 
-    if (await _repoManager.Address.DeleteAddress(request.AddressID, request.UserID))
+    if (await _addressDbContext.DeleteAddress(request.AddressID, request.UserID))
       return Ok();
 
     return StatusCode(StatusCodes.Status502BadGateway);
