+++ Infrastructure/Common/SecurityUtil.cs
