+++ Infrastructure/Persistence/Configurations/ProductConfConfigurations.cs
@@ -2,13 +2,12 @@ using Domain.Entities.Product;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class ProductConfConfigurations : IEntityTypeConfiguration<ProductConf>
 {
-    public class ProductConfConfigurations : IEntityTypeConfiguration<ProductConf>
-  {
     public void Configure(EntityTypeBuilder<ProductConf> builder)
     {
-      builder.HasKey(x => new { x.ProductItemID, x.VariationOptionID });
+        builder.HasKey(x => new { x.ProductItemID, x.VariationOptionID });
     }
-  }
 }
