+++ Infrastructure/Persistence/Configurations/PromotionCategoriesConfigurations.cs
@@ -2,13 +2,12 @@ using Domain.Entities.Product;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class PromotionCategoriesConfigurations : IEntityTypeConfiguration<PromotionCategories>
 {
-    public class PromotionCategoriesConfigurations : IEntityTypeConfiguration<PromotionCategories>
-  {
     public void Configure(EntityTypeBuilder<PromotionCategories> builder)
     {
-      builder.HasKey(x => new { x.CategoryID, x.PromotionID });
+        builder.HasKey(x => new { x.CategoryID, x.PromotionID });
     }
-  }
 }
