+++ Infrastructure/Persistence/Configurations/PromotionProductsConfigurations.cs
@@ -2,13 +2,12 @@ using Domain.Entities.Product;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class PromotionProductsConfigurations : IEntityTypeConfiguration<PromotionProducts>
 {
-    public class PromotionProductsConfigurations : IEntityTypeConfiguration<PromotionProducts>
-  {
     public void Configure(EntityTypeBuilder<PromotionProducts> builder)
     {
-      builder.HasKey(x => new { x.ProductItemID, x.PromotionID });
+        builder.HasKey(x => new { x.ProductItemID, x.PromotionID });
     }
-  }
 }
