+++ Infrastructure/Persistence/Configurations/RoleConfigurations.cs
@@ -3,19 +3,18 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using static Domain.Entities.Auth.Role;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class RoleConfigurations : IEntityTypeConfiguration<Role>
 {
-    public class RoleConfigurations : IEntityTypeConfiguration<Role>
-  {
     public void Configure(EntityTypeBuilder<Role> builder)
     {
-      builder.HasData(
-       Enum.GetValues(typeof(UserRoles))
-         .Cast<UserRoles>().Select(u => new Role()
-         {
-           ID = u,
-           Name = u.ToString(),
-         }));
+        builder.HasData(
+         Enum.GetValues(typeof(UserRoles))
+           .Cast<UserRoles>().Select(u => new Role()
+           {
+               ID = u,
+               Name = u.ToString(),
+           }));
     }
-  }
 }
