+++ Infrastructure/Persistence/Configurations/UserAddressConfigurations.cs
@@ -2,14 +2,12 @@ using Domain.Entities.User;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class UserAddressConfigurations : IEntityTypeConfiguration<UserAddress>
 {
-    public class UserAddressConfigurations : IEntityTypeConfiguration<UserAddress>
-  {
     public void Configure(EntityTypeBuilder<UserAddress> builder)
     {
-      builder.HasKey(x => new { x.UserID, x.AddressID });
+        builder.HasKey(x => new { x.UserID, x.AddressID });
     }
-  }
-
 }
