+++ Infrastructure/Persistence/Configurations/UserAuthMethodConfigurations.cs
@@ -2,13 +2,12 @@ using Domain.Entities.User;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class UserAuthMethodConfigurations : IEntityTypeConfiguration<UserAuthMethod>
 {
-    public class UserAuthMethodConfigurations : IEntityTypeConfiguration<UserAuthMethod>
-  {
     public void Configure(EntityTypeBuilder<UserAuthMethod> builder)
     {
-      builder.HasKey(x => new { x.UserID, x.AuthProviderID });
+        builder.HasKey(x => new { x.UserID, x.AuthProviderID });
     }
-  }
 }
