+++ Infrastructure/Persistence/Configurations/UserVerificationMethodConfigurations.cs
@@ -2,13 +2,12 @@ using Domain.Entities.User;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class UserVerificationMethodConfigurations : IEntityTypeConfiguration<UserVerificationMethod>
 {
-    public class UserVerificationMethodConfigurations : IEntityTypeConfiguration<UserVerificationMethod>
-  {
     public void Configure(EntityTypeBuilder<UserVerificationMethod> builder)
     {
-      builder.HasKey(x => new { x.UserID, x.VerificationMethodID });
+        builder.HasKey(x => new { x.UserID, x.VerificationMethodID });
     }
-  }
 }
