+++ Infrastructure/Persistence/Configurations/VerificationMethodConfigurations.cs
@@ -3,19 +3,18 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using static Domain.Entities.Auth.VerificationMethod;
 
-namespace Shop.RepoConfigurations
+namespace Infrastructure.Persistence.Configurations;
+
+public class VerificationMethodConfigurations : IEntityTypeConfiguration<VerificationMethod>
 {
-    public class VerificationMethodConfigurations : IEntityTypeConfiguration<VerificationMethod>
-  {
     public void Configure(EntityTypeBuilder<VerificationMethod> builder)
     {
-      builder.HasData(
-      Enum.GetValues(typeof(VerifyMethods))
-      .Cast<VerifyMethods>().Select(u => new VerificationMethod()
-      {
-        ID = u,
-        Name = u.ToString()
-      }));
+        builder.HasData(
+        Enum.GetValues(typeof(VerifyMethods))
+        .Cast<VerifyMethods>().Select(u => new VerificationMethod()
+        {
+            ID = u,
+            Name = u.ToString()
+        }));
     }
-  }
 }
